Implementación auth
Implementación parcial de auth

+++ src/main/resources/db/migration/V1__DDL.sql
@@ -2,49 +2,49 @@ CREATE TYPE rol_usuario AS ENUM (
     'cliente',
     'operador',
     'administrador'
-);
+    );
 
 CREATE TYPE reserva_estado AS ENUM (
     'reservado',
     'en_curso',
     'cancelado',
     'completado'
-);
+    );
 
 CREATE TYPE salida_estado AS ENUM (
     'programado',
     'en_curso',
     'completado',
     'cancelado'
-);
+    );
 
 CREATE TYPE nivel_dificultad AS ENUM (
     'facil',
     'moderado',
     'dificil'
-);
+    );
 
 CREATE TYPE tipo_documento AS ENUM (
     'cedula',
     'pasaporte',
     'tarjeta_identidad'
-);
+    );
 
 -- =============================================================
 --  TABLA: usuarios
 -- =============================================================
 
 CREATE TABLE usuarios (
-    id                  SERIAL          PRIMARY KEY,
-    primer_nombre       VARCHAR(100)    NOT NULL,
-    primer_apellido     VARCHAR(100)    NOT NULL,
-    tipo_documento      tipo_documento  NOT NULL,
-    documento           VARCHAR(50)     NOT NULL,
-    email               VARCHAR(200)    NOT NULL UNIQUE,
-    password_hash       VARCHAR(255)    NOT NULL,
-    telefono            VARCHAR(20),
-    role                rol_usuario     NOT NULL DEFAULT 'cliente',
-    is_active           BOOLEAN         NOT NULL DEFAULT TRUE
+                          id                  BIGSERIAL       PRIMARY KEY,
+                          primer_nombre       VARCHAR(100)    NOT NULL,
+                          primer_apellido     VARCHAR(100)    NOT NULL,
+                          tipo_documento      tipo_documento  NOT NULL,
+                          documento           VARCHAR(50)     NOT NULL,
+                          email               VARCHAR(200)    NOT NULL UNIQUE,
+                          password_hash       VARCHAR(255)    NOT NULL,
+                          telefono            VARCHAR(20),
+                          role                rol_usuario     NOT NULL DEFAULT 'cliente',
+                          is_active           BOOLEAN         NOT NULL DEFAULT TRUE
 );
 
 COMMENT ON TABLE  usuarios               IS 'Usuarios del sistema: clientes, operadores y administradores.';
@@ -55,15 +55,15 @@ COMMENT ON COLUMN usuarios.password_hash IS 'Contraseña almacenada con hash.';
 -- =============================================================
 
 CREATE TABLE rutas (
-    id                  SERIAL              PRIMARY KEY,
-    nombre              VARCHAR(150)        NOT NULL,
-    descripcion         TEXT,
-    dificultad          nivel_dificultad    NOT NULL,
-    duracion_minutos    INT                 NOT NULL CHECK (duracion_minutos > 0),
-    max_caballos        INT                 NOT NULL CHECK (max_caballos > 0),
-    min_guias           INT                 NOT NULL DEFAULT 1 CHECK (min_guias > 0),
-    image_url           VARCHAR(500),
-    is_active           BOOLEAN             NOT NULL DEFAULT TRUE
+                       id                  BIGSERIAL           PRIMARY KEY,
+                       nombre              VARCHAR(150)        NOT NULL,
+                       descripcion         TEXT,
+                       dificultad          nivel_dificultad    NOT NULL,
+                       duracion_minutos    INT                 NOT NULL CHECK (duracion_minutos > 0),
+                       max_caballos        INT                 NOT NULL CHECK (max_caballos > 0),
+                       min_guias           INT                 NOT NULL DEFAULT 1 CHECK (min_guias > 0),
+                       image_url           VARCHAR(500),
+                       is_active           BOOLEAN             NOT NULL DEFAULT TRUE
 );
 
 COMMENT ON TABLE  rutas                  IS 'Rutas turísticas. Solo los administradores pueden crearlas.';
@@ -77,7 +77,7 @@ COMMENT ON COLUMN rutas.duracion_minutos IS 'Duración estimada en minutos; dete
 
 CREATE TABLE caballos
 (
-    id          SERIAL PRIMARY KEY,
+    id          BIGSERIAL       PRIMARY KEY,
     nombre      VARCHAR(100)    NOT NULL,
     raza        VARCHAR(100),
     is_active   BOOLEAN         NOT NULL DEFAULT TRUE
@@ -90,11 +90,11 @@ COMMENT ON TABLE caballos IS 'Recurso: caballos disponibles para las rutas.';
 -- =============================================================
 
 CREATE TABLE guias (
-    id          SERIAL          PRIMARY KEY,
-    nombre      VARCHAR(100)    NOT NULL,
-    telefono    VARCHAR(20),
-    email       VARCHAR(150),
-    is_active   BOOLEAN         NOT NULL DEFAULT TRUE
+                       id          BIGSERIAL       PRIMARY KEY,
+                       nombre      VARCHAR(100)    NOT NULL,
+                       telefono    VARCHAR(20),
+                       email       VARCHAR(150),
+                       is_active   BOOLEAN         NOT NULL DEFAULT TRUE
 );
 
 COMMENT ON TABLE guias IS 'Recurso: guías turísticos.';
@@ -107,13 +107,13 @@ COMMENT ON TABLE guias IS 'Recurso: guías turísticos.';
 -- =============================================================
 
 CREATE TABLE salidas (
-    id                  SERIAL          PRIMARY KEY,
-    ruta_id             INT             NOT NULL REFERENCES rutas (id),
-    fecha_programada    DATE            NOT NULL,
-    tiempo_inicio       TIME            NOT NULL,
-    tiempo_fin          TIME            NOT NULL,
-    estado              salida_estado   NOT NULL DEFAULT 'programado',
-    CONSTRAINT chk_salida_tipo CHECK (tiempo_fin > tiempo_inicio)
+                         id                  BIGSERIAL       PRIMARY KEY,
+                         ruta_id             BIGINT          NOT NULL REFERENCES rutas (id),
+                         fecha_programada    DATE            NOT NULL,
+                         tiempo_inicio       TIME            NOT NULL,
+                         tiempo_fin          TIME            NOT NULL,
+                         estado              salida_estado   NOT NULL DEFAULT 'programado',
+                         CONSTRAINT chk_salida_tipo CHECK (tiempo_fin > tiempo_inicio)
 );
 
 COMMENT ON TABLE  salidas          IS 'Salida concreta (ruta + fecha + hora). Agrupa reservas que viajan juntas.';
@@ -125,10 +125,10 @@ COMMENT ON COLUMN salidas.tiempo_fin IS 'tiempo_inicio + duration_minutes de la 
 -- =============================================================
 
 CREATE TABLE salida_caballos (
-    id          SERIAL  PRIMARY KEY,
-    salida_id   INT     NOT NULL REFERENCES salidas (id) ON DELETE CASCADE,
-    horse_id    INT     NOT NULL REFERENCES caballos (id),
-    CONSTRAINT uq_salida_caballo UNIQUE (salida_id, horse_id)
+                                 id          BIGSERIAL   PRIMARY KEY,
+                                 salida_id   BIGINT      NOT NULL REFERENCES salidas (id) ON DELETE CASCADE,
+                                 horse_id    BIGINT      NOT NULL REFERENCES caballos (id),
+                                 CONSTRAINT uq_salida_caballo UNIQUE (salida_id, horse_id)
 );
 
 COMMENT ON TABLE salida_caballos IS 'Caballos asignados a una salida. Cantidad <= routes.max_horses.';
@@ -139,10 +139,10 @@ COMMENT ON TABLE salida_caballos IS 'Caballos asignados a una salida. Cantidad <
 -- =============================================================
 
 CREATE TABLE salida_guias (
-    id          SERIAL PRIMARY KEY,
-    salida_id   INT    NOT NULL REFERENCES salidas (id) ON DELETE CASCADE,
-    guia_id     INT    NOT NULL REFERENCES guias (id),
-    CONSTRAINT uq_salida_guia UNIQUE (salida_id, guia_id)
+                              id          BIGSERIAL   PRIMARY KEY,
+                              salida_id   BIGINT      NOT NULL REFERENCES salidas (id) ON DELETE CASCADE,
+                              guia_id     BIGINT      NOT NULL REFERENCES guias (id),
+                              CONSTRAINT uq_salida_guia UNIQUE (salida_id, guia_id)
 );
 
 COMMENT ON TABLE salida_guias IS 'Guías asignados a una salida. Cantidad >= routes.min_guides.';
@@ -152,12 +152,12 @@ COMMENT ON TABLE salida_guias IS 'Guías asignados a una salida. Cantidad >= rou
 -- =============================================================
 
 CREATE TABLE reservaciones (
-    id            SERIAL             PRIMARY KEY,
-    salida_id     INT                NOT NULL REFERENCES salidas (id),
-    client_id     INT                NOT NULL REFERENCES usuarios (id),
-    operator_id   INT                REFERENCES usuarios (id),
-    num_people    INT                NOT NULL DEFAULT 1 CHECK (num_people > 0),
-    estado        reserva_estado NOT NULL DEFAULT 'reservado'
+                               id            BIGSERIAL          PRIMARY KEY,
+                               salida_id     BIGINT             NOT NULL REFERENCES salidas (id),
+                               client_id     BIGINT             NOT NULL REFERENCES usuarios (id),
+                               operator_id   BIGINT             REFERENCES usuarios (id),
+                               num_people    INT                NOT NULL DEFAULT 1 CHECK (num_people > 0),
+                               estado        reserva_estado     NOT NULL DEFAULT 'reservado'
 );
 
 COMMENT ON TABLE  reservaciones             IS 'Reserva de un grupo para una salida específica.';
@@ -171,20 +171,22 @@ COMMENT ON COLUMN reservaciones.num_people  IS 'Cantidad de personas en el grupo
 -- =============================================================
 
 CREATE TABLE participantes (
-    id                  SERIAL          PRIMARY KEY,
-    reservacion_id      INT             NOT NULL REFERENCES reservaciones (id) ON DELETE CASCADE,
-    primer_nombre       VARCHAR(100)    NOT NULL,
-    primer_apellido     VARCHAR(100)    NOT NULL,
-    tipo_documento      tipo_documento  NOT NULL,
-    documento           VARCHAR(50)     NOT NULL,
-    edad                SMALLINT        NOT NULL CHECK (edad > 0 AND edad < 120),
-    altura_cm           SMALLINT        NOT NULL CHECK (altura_cm > 0),
-    peso_kg             NUMERIC(5,2)    NOT NULL CHECK (peso_kg > 0),
+                               id                  BIGSERIAL       PRIMARY KEY,
+                               reservacion_id      BIGINT          NOT NULL REFERENCES reservaciones (id) ON DELETE CASCADE,
+                               primer_nombre       VARCHAR(100)    NOT NULL,
+                               primer_apellido     VARCHAR(100)    NOT NULL,
+                               tipo_documento      tipo_documento  NOT NULL,
+                               documento           VARCHAR(50)     NOT NULL,
+                               edad                SMALLINT        NOT NULL CHECK (edad > 0 AND edad < 120),
+                               altura_cm           SMALLINT        NOT NULL CHECK (altura_cm > 0),
+                               peso_kg             NUMERIC(5,2)    NOT NULL CHECK (peso_kg > 0),
 
-    CONSTRAINT uq_participant_doc UNIQUE (reservacion_id, tipo_documento, documento)
+                               CONSTRAINT uq_participant_doc UNIQUE (reservacion_id, tipo_documento, documento)
 );
 
 COMMENT ON TABLE  participantes                 IS 'Datos personales de cada integrante de una reserva.';
 COMMENT ON COLUMN participantes.altura_cm       IS 'Estatura en centímetros (ej. 170).';
 COMMENT ON COLUMN participantes.peso_kg       IS 'Peso en kilogramos con hasta 2 decimales (ej. 75.50).';
 COMMENT ON COLUMN participantes.documento IS 'Número de cédula, pasaporte o tarjeta de identidad.';
+
+
